Backend/requirements.txt
streamlit
tensorflow
pillow
pandas
flask